Se agregaron librerias y se corrigieron errores al agregar un producto al carrito y poder eliminar

+++ js/main.js
@@ -1,114 +1,17 @@
-// alert("Bienvenido a FarmApp - Tu farmacia de confianza \n\nNuestra plataforma esta en mantenimiento 🛠️\npor ahora solo puedes comprar 1 producto a la vez. ");
-// let producto;
-// let cantidad;
-// let total;
-// let seguirComprando = true;
-// let subTotal = function (a, b) {
-//     switch (a) {
-//         case 1:
-//             // Aspirinas
-//             return 1 * b;
-//         case 2:
-//             // Panadol Ultra
-//             return 2 * b;
-//         case 3:
-//             // Bioelectro
-//             return 4 * b;
-//         case 4:
-//             // Tabcin gripe y tos
-//             return 3 * b;
-//     }
-// }
-// do {
-//     producto = parseInt(prompt("Productos: \n 1.- Aspirina $1 \n 2.- Panadol Ultra $2 \n 3.- Bioelectro $4 \n 4.- Tabcin gripe y tos $3"));
-//     if (producto < 5 && producto > 0) {
-//         cantidad = parseInt(prompt("Cuantos deseas comprar: "));
-
-//         if (!isNaN(producto) && !isNaN(cantidad)) {
-//             // si es número
-//             total = subTotal(producto, cantidad);
-//             alert(`El total a pagar es: ${total}`)
-//         } else {
-//             alert(`porfavor ingrese valores correctos`);
-//         }
-//         seguirComprando = confirm(`¿Deseas comprar otro producto?`);
-//         console.log(seguirComprando);
-//     } else {
-//         alert(`Porfavor Ingresa un valor en el rango de 1 - 4`)
-//     }
-
-// } while (seguirComprando)
-
-// -------------------------------------------------------------
-//               SEGUNDA PRE-ENTREGA (Refactorización)
-// -------------------------------------------------------------
-
-// const productos = [
-//     {
-//         nombre: "Panadol Ultra",
-//         precio: 2,
-//     },
-//     {
-//         nombre: "Aspirina",
-//         precio: 1,
-//     },
-//     {
-//         nombre: "BioElectro",
-//         precio: 3,
-//     },
-//     {
-//         nombre: "Tabcin Gripe y Tos",
-//         precio: 2,
-//     }
-// ]
-
-// function mostrarProductos(){
-//     let mensaje='';
-//     let cntTitulo=0;
-//     let noProducto = 0;
-//     productos.forEach(el => {
-//         noProducto++
-//         if(cntTitulo===0){
-//             mensaje += 'Productos: \n';
-//             cntTitulo++;
-//         }
-//         mensaje += noProducto + ". " + el.nombre + " " +"$" + el.precio + "\n"
-//     });
-//     return mensaje;
-// }
-
-// let seleccionado = (el,cntd) =>{
-//     let subtotal = productos[el].precio*cntd;
-//     alert(`El total de su compra es: $${subtotal}`);
-// }
-
-// let seguirComprando = true;
-
-// do{
-
-//     let producto = parseInt(prompt(mostrarProductos()));
-    
-    
-//     if(producto>0 && producto<=productos.length){
-//         producto--;
-//         let cantidad = parseInt(prompt("¿Cuantos desea comprar?"));
-//         seleccionado(producto, cantidad);
-//     }else{
-//         alert("No existe el producto :(");
-//     }
-
-//     seguirComprando = confirm('¿Quieres Seguir comprando?');
-// }while(seguirComprando)
-
-// --------------------------------------------------------------
-//              TERCERA PREENTREGA - DOM 
-// --------------------------------------------------------------
-
 let carrito = JSON.parse(localStorage.getItem("carrito")) || [];
 const botones = document.getElementById("button__header");
 const container = document.getElementById("container");
 
-function crearCard(arrayMedicamentos){
+let mostrarOcultar = false;
+
+function eliminarDelCarrito(nombre) {
+    let nuevoCarrito = carrito.filter(el => el.nombre !== nombre);
+    carrito = nuevoCarrito;
+    localStorage.setItem("carrito", JSON.stringify(carrito));
+    mostrarCarrito(carrito);
+}
+
+function crearCard(arrayMedicamentos) {
 
     const card = document.createElement("div");
     card.className = "card"
@@ -120,45 +23,76 @@ function crearCard(arrayMedicamentos){
     const tituloMedicamento = document.createElement("p");
     tituloMedicamento.innerText = arrayMedicamentos.nombre;
     tituloMedicamento.className = "tituloMedicamento";
-    
+
 
     const precioMedicamento = document.createElement("p");
     precioMedicamento.innerText = `$${arrayMedicamentos.precio}`;
     precioMedicamento.className = "precioMedicamento";
-    
-    const agregarAlCarrito = document.createElement("button");
-    agregarAlCarrito.innerText = "Agregar";
-    agregarAlCarrito.onclick = () =>{
-        carrito.push(arrayMedicamentos);
-        console.log(carrito);
-        localStorage.setItem("carrito", JSON.stringify(carrito));
-    };
 
+    const agregarAlCarrito = document.createElement("button");
+    agregarAlCarrito.innerText = mostrarOcultar ? "Eliminar" : "Agregar";
+
+    if (mostrarOcultar) {
+        agregarAlCarrito.onclick = () => {
+            eliminarDelCarrito(arrayMedicamentos.nombre);
+        };
+    } else {
+        agregarAlCarrito.onclick = () => {
+            
+            const productoExistente = carrito.find(el => el.nombre === arrayMedicamentos.nombre);
+            if (!productoExistente) {
+                carrito.push(arrayMedicamentos);
+                localStorage.setItem("carrito", JSON.stringify(carrito));
+                Swal.fire({
+                    title: "AGREGADO",
+                    text: `Se añadió ${arrayMedicamentos.nombre} al carrito`,
+                    icon: "success"
+                });
+            } else {
+                Swal.fire({
+                    title: "Ya está en el carrito",
+                    text: `${arrayMedicamentos.nombre} ya está en el carrito`,
+                    icon: "info"
+                });
+            }
+        };
+    }
     card.append(img, tituloMedicamento, precioMedicamento, agregarAlCarrito);
     container.append(card);
 }
 
 
-function mostrarMedicamentos(){
-    container.innerText="";
+function mostrarMedicamentos() {
+    container.innerText = "";
+    mostrarOcultar = false;
+
     medicamentos.forEach(el => {
         crearCard(el)
     });
 }
 
-function mostrarCarrito(arrayCarrito){
-    container.innerText="";
+function mostrarCarrito(arrayCarrito) {
+    container.innerHTML= "";
+    mostrarOcultar = true;
     arrayCarrito.forEach(el => {
         crearCard(el);
     });
 }
 
-if(container.innerText != "" ){
-    console.log("El container no esta vacío");
-}else{
-    
-    medicamentos.forEach(el => {
-        crearCard(el)
+if (container.innerHTML != "") {
+    setTimeout(() => {
+        const loader = document.getElementById("loader");
+        loader.remove();
+        // console.log("estoy aca");
+        medicamentos.forEach(el => {
+            crearCard(el)
+        });
+    }, 2000)
+} else {
+    Swal.fire({
+        title: "OOPS",
+        text: `A OCURRIDO UN ERROR AL CARGAR LOS PRODUCTOS`,
+        icon: "error"
     });
 }
 
@@ -174,7 +108,5 @@ btnCarrito.className = "btnProductos font--button--poppins--600"
 botones.appendChild(btnProductos);
 botones.appendChild(btnCarrito);
 
-btnProductos.onclick= () => mostrarMedicamentos();
+btnProductos.onclick = () => mostrarMedicamentos();
 btnCarrito.onclick = () => mostrarCarrito(carrito);
-
-
